Fix a race condition in vmaf_thread_pool_wait.
There may still be work in the queue but n_working=0 before the vmaf_thread_pool_runner loop fetches another job in the queue.  When calling vmaf_thread_pool_wait (and not stopping), we will still want to drain the queue. So only exit the wait loop when n_working = 0 && head == nullptr